Incorporate hilt-binder library
