+++ app/build.gradle.kts
@@ -56,10 +56,19 @@ android {
         viewBinding = true
     }
 
+    // https://dagger.dev/hilt/gradle-setup#classpath-aggregation
+    lintOptions {
+        isCheckReleaseBuilds = false
+    }
+
     compileOptions {
         sourceCompatibility = appConfig.javaCompatibilityVersion
         targetCompatibility = appConfig.javaCompatibilityVersion
     }
+}
+
+hilt {
+    enableExperimentalClasspathAggregation = true
 }
 
 dependencies {
@@ -104,6 +113,9 @@ dependencies {
 
     implementation(deps.androidX.daggerHiltAssistedInjection)
     kapt(deps.androidX.daggerHiltAssistedInjectionCompiler)
+
+    implementation(deps.misc.hiltBinder)
+    kapt(deps.misc.hiltBinderCompiler)
 
     testImplementation(deps.testing.jUnit)
     androidTestImplementation(deps.testing.jUnitExt)
