--- core/src/main/java/com/paulrybitskyi/docskanner/core/di/FormattersModule.kt
@@ -1,38 +1,0 @@
-/*
- * Copyright 2020 Paul Rybitskyi, paul.rybitskyi.work@gmail.com
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.paulrybitskyi.docskanner.core.di
-
-import com.paulrybitskyi.docskanner.core.formatters.DocDateFormatter
-import com.paulrybitskyi.docskanner.core.formatters.DocDateFormatterImpl
-import com.paulrybitskyi.docskanner.core.formatters.DocSizeFormatter
-import com.paulrybitskyi.docskanner.core.formatters.DocSizeFormatterImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal interface FormattersModule {
-
-    @Binds
-    fun bindDocDateFormatter(formatter: DocDateFormatterImpl): DocDateFormatter
-
-    @Binds
-    fun bindDocSizeFormatter(formatter: DocSizeFormatterImpl): DocSizeFormatter
-
-}
