+++ core/src/main/java/com/paulrybitskyi/docskanner/core/providers/StringProvider.kt
@@ -17,6 +17,7 @@
 package com.paulrybitskyi.docskanner.core.providers
 
 import android.content.Context
+import com.paulrybitskyi.hiltbinder.BindType
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
 
@@ -30,6 +31,7 @@ interface StringProvider {
 }
 
 
+@BindType
 internal class StringProviderImpl @Inject constructor(
     @ApplicationContext private val applicationContext: Context
 ) : StringProvider {
