+++ data/src/main/java/com/paulrybitskyi/docskanner/data/ClearAppCacheUseCaseImpl.kt
@@ -14,11 +14,14 @@
  * limitations under the License.
  */
 
-package com.paulrybitskyi.docskanner.data.usecases
+package com.paulrybitskyi.docskanner.data
 
-import com.paulrybitskyi.docskanner.core.providers.AppStorageFolderProvider
+import android.content.Context
+import com.paulrybitskyi.commons.ktx.fileList
 import com.paulrybitskyi.docskanner.core.providers.DispatcherProvider
-import com.paulrybitskyi.docskanner.domain.CreateAppStorageFolderUseCase
+import com.paulrybitskyi.docskanner.domain.ClearAppCacheUseCase
+import com.paulrybitskyi.hiltbinder.BindType
+import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
@@ -26,22 +29,20 @@ import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-internal class CreateAppStorageFolderUseCaseImpl @Inject constructor(
-    private val appStorageFolderProvider: AppStorageFolderProvider,
+@BindType
+internal class ClearAppCacheUseCaseImpl @Inject constructor(
+    @ApplicationContext private val applicationContext: Context,
     private val dispatcherProvider: DispatcherProvider
-) : CreateAppStorageFolderUseCase {
+) : ClearAppCacheUseCase {
 
 
     override suspend fun execute(params: Unit): Flow<Unit> {
         return flow<Unit> {
-            val appStorageFolder = appStorageFolderProvider.getAppStorageFolder()
+            val cacheFolder = applicationContext.cacheDir
+            val cacheFiles = cacheFolder.fileList()
 
-            if(!appStorageFolder.exists()) {
-                appStorageFolder.mkdirs()
-            }
-
-            if(!appStorageFolder.exists()) {
-                throw IllegalStateException("Could not create the app's storage folder.")
+            for(cacheFile in cacheFiles) {
+                if(cacheFile.exists()) cacheFile.delete()
             }
         }
         .flowOn(dispatcherProvider.io)
