--- image-loading/src/main/java/com/paulrybitskyi/docskanner/imageloading/di/ImageLoadingBindingsModule.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright 2020 Paul Rybitskyi, paul.rybitskyi.work@gmail.com
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.paulrybitskyi.docskanner.imageloading.di
-
-import com.paulrybitskyi.docskanner.imageloading.ImageLoader
-import com.paulrybitskyi.docskanner.imageloading.ImageLoaderImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal interface ImageLoadingBindingsModule {
-
-    @Binds
-    fun bindImageLoader(imageLoader: ImageLoaderImpl): ImageLoader
-
-}
